backend/chatbot/requirements.txt
# For Gemini integration
google-generativeai==0.3.1
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
# Add your other requirements